third_party/blink/web_tests/virtual/disable-single-axis-scroll-containers/external/wpt/css/css-position/sticky/position-sticky-single-axis-used-values-expected.txt

This is a testharness.js-based test.
[FAIL] The used values from getComputedStyle() resolve against the correct scroller.
  assert_equals: expected "75px" but got "40px"
Harness: the test ran to completion.
